+++ etc/dbus-serialbattery/utils.py
@@ -36,7 +36,7 @@ def _get_list_from_config(
 
 # Constants - Need to dynamically get them in future
 DRIVER_VERSION = "1.0"
-DRIVER_SUBVERSION = ".0 (20230427)"
+DRIVER_SUBVERSION = ".0 (20230428)"
 zero_char = chr(48)
 degree_sign = "\N{DEGREE SIGN}"